#DL-improvment implements condition for drawing lines

+++ src/com/library/drawing/draw/controller/DrawController.java
@@ -24,6 +24,10 @@ public class DrawController {
 
     List<String> rectangles = new ArrayList<>();
 
+    List<String> lines = new ArrayList<>();
+    Integer width;
+    Integer height;
+
     String userCommand;
 
     /**
@@ -39,8 +43,8 @@ public class DrawController {
         char firstCharacter;
         try {
             if (arrayLength == 3 && (character == 'c' || character == 'C')) {
-                Integer width = Integer.parseInt(inputString[1]);
-                Integer height = Integer.parseInt(inputString[2]);
+                width = Integer.parseInt(inputString[1]);
+                height = Integer.parseInt(inputString[2]);
                 canvas = drawServiceImpl.drawCanvas(height, width);
                 drawServiceImpl.display(canvas);
                 do {
@@ -106,13 +110,21 @@ public class DrawController {
         Integer y1 = Integer.parseInt(command[2]);
         Integer x2 = Integer.parseInt(command[3]);
         Integer y2 = Integer.parseInt(command[4]);
-        if (x1 >= 0 && x2 <= width && y1 >= 0 && y2 <= height) {
-            Map<String, Integer> coordinates = drawHelper.getMaxMin(x1, y1, x2, y2);
+        Map<String, Integer> coordinates = drawHelper.getMaxMin(x1, y1, x2, y2);
+        Integer minX = coordinates.get("minX");
+        Integer minY = coordinates.get("minY");
+        Integer maxX = coordinates.get("maxX");
+        Integer maxY = coordinates.get("maxY");
+        if (((minX == 0 || maxX == width - 1) && (minY == 0 || maxY == height - 1))
+                || ((maxY == height - 1 || minY == 0) && (minX == 0 || maxX == width - 1))) {
+            lines.add(userCommand);
             canvas = drawServiceImpl.drawLine(canvas, coordinates);
-            drawServiceImpl.display(canvas);
-        } else {
+        } else if (minX < 0 || maxX >= width || minY < 0 || maxY >= height) {
             System.out.println("Coordinates are outside the canvas boundary.");
+        } else {
+            canvas = drawServiceImpl.line(canvas, coordinates);
         }
+        drawServiceImpl.display(canvas);
     }
 
     /**
@@ -125,16 +137,45 @@ public class DrawController {
         Integer x = Integer.parseInt(command[1]);
         Integer y = Integer.parseInt(command[2]);
         char c = command[3].charAt(0);
-        canvas = drawServiceImpl.collision(canvas, y, x, c);
-        for (String rectangle : rectangles) {
-            String[] splitCommand = drawHelper.splitWhitespace(rectangle);
+        drawBubble(rectangles, x, y, c);
+        drawBubble(lines, x, y, c);
+        /*
+         * canvas = drawServiceImpl.collision(canvas, y, x, c); canvas = drawServiceImpl.fillAll(canvas, c); for (String rectangle :
+         * rectangles) { String[] splitCommand = drawHelper.splitWhitespace(rectangle); Integer x1 = Integer.parseInt(splitCommand[1]);
+         * Integer y1 = Integer.parseInt(splitCommand[2]); Integer x2 = Integer.parseInt(splitCommand[3]); Integer y2 =
+         * Integer.parseInt(splitCommand[4]); canvas = drawServiceImpl.removeFill(canvas, x1, y1, x2, y2); }
+         * 
+         * canvas = drawServiceImpl.removeFill(canvas, 6, 0, 11, 3);
+         */
+
+    }
+
+    public void drawBubble(List<String> figures, int x, int y, char c) {
+        Map<String, Integer> coordinates;
+        for (String figure : figures) {
+            String[] splitCommand = drawHelper.splitWhitespace(figure);
+            char firstCharacter = splitCommand[0].charAt(0);
             Integer x1 = Integer.parseInt(splitCommand[1]);
             Integer y1 = Integer.parseInt(splitCommand[2]);
             Integer x2 = Integer.parseInt(splitCommand[3]);
             Integer y2 = Integer.parseInt(splitCommand[4]);
-            canvas = drawServiceImpl.removeFill(canvas, x1, y1, x2, y2);
+            coordinates = drawHelper.getMaxMin(x1, y1, x2, y2);
+            if (Character.toLowerCase(firstCharacter) == 'r') {
+                Boolean isInsideRectangle = drawServiceImpl.isInsideRectangle(x, y, coordinates);
+                if (isInsideRectangle) {
+                    canvas = drawServiceImpl.collision(canvas, y, x, c);
+
+                } else {
+                    canvas = drawServiceImpl.fillAll(canvas, c);
+                    canvas = drawServiceImpl.removeFill(canvas, x1, y1, x2, y2);
+                }
+
+            } else if (Character.toLowerCase(firstCharacter) == 'l') {
+                canvas = drawServiceImpl.fillAll(canvas, c);
+                canvas = drawServiceImpl.removeLineFill(canvas, x1, y1, x2, y2, width, height);
+            }
+            drawServiceImpl.display(canvas);
         }
-        drawServiceImpl.display(canvas);
     }
 
 }
